replace orm model to basic domain
